Improve the "replaceRomanNumerals" config options.
This extends support to potion potency levels as requested in #2. This
also shuffles mixins around to take advantage of new mixin capabilities.

The Component mixin can now be applied on NeoForge and has been copied to
the NeoForge project sources. It can't live in common because Forge does
not support interface injectors and their annotation processor will error
even if they are not loaded in the Forge mixin config.

This change also implements some small changes to the safety of the
original code. We check if the potency/level is actually a number before
modifying the value. This is likely unneeded but it's possible for another
mod to implement a similarly formatted localization key that would break
our code.